refactor: move provider settings use cases
