+++ internal/api/handlers/management/config_lists.go
@@ -638,7 +638,7 @@ func (h *Handler) DeleteClaudeKey(c *gin.Context) {
 
 // bedrock-api-key: []BedrockKey
 func (h *Handler) GetBedrockKeys(c *gin.Context) {
-	c.JSON(200, gin.H{"bedrock-api-key": h.cfg.BedrockKey})
+	c.JSON(200, gin.H{"bedrock-api-key": providerSettingsService(h).BedrockKeys()})
 }
 
 func (h *Handler) PutBedrockKeys(c *gin.Context) {
@@ -658,14 +658,7 @@ func (h *Handler) PutBedrockKeys(c *gin.Context) {
 		}
 		arr = obj.Items
 	}
-	for i := range arr {
-		normalizeBedrockKey(&arr[i])
-	}
-	prev := append([]config.BedrockKey(nil), h.cfg.BedrockKey...)
-	h.cfg.BedrockKey = arr
-	h.cfg.SanitizeBedrockKeys()
-	if err := h.validateChannelNames(); err != nil {
-		h.cfg.BedrockKey = prev
+	if err := providerSettingsService(h).ReplaceBedrockKeys(arr); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
@@ -673,107 +666,20 @@ func (h *Handler) PutBedrockKeys(c *gin.Context) {
 }
 
 func (h *Handler) PatchBedrockKey(c *gin.Context) {
-	type bedrockKeyPatch struct {
-		Name            *string                `json:"name"`
-		Priority        *int                   `json:"priority"`
-		Prefix          *string                `json:"prefix"`
-		AuthMode        *string                `json:"auth-mode"`
-		APIKey          *string                `json:"api-key"`
-		AccessKeyID     *string                `json:"access-key-id"`
-		SecretAccessKey *string                `json:"secret-access-key"`
-		SessionToken    *string                `json:"session-token"`
-		Region          *string                `json:"region"`
-		ForceGlobal     *bool                  `json:"force-global"`
-		BaseURL         *string                `json:"base-url"`
-		ProxyURL        *string                `json:"proxy-url"`
-		ProxyID         *string                `json:"proxy-id"`
-		Models          *[]config.BedrockModel `json:"models"`
-		Headers         *map[string]string     `json:"headers"`
-		ExcludedModels  *[]string              `json:"excluded-models"`
-	}
 	var body struct {
-		Index *int             `json:"index"`
-		Match *string          `json:"match"`
-		Value *bedrockKeyPatch `json:"value"`
+		Index *int                              `json:"index"`
+		Match *string                           `json:"match"`
+		Value *providersettings.BedrockKeyPatch `json:"value"`
 	}
 	if err := c.ShouldBindJSON(&body); err != nil || body.Value == nil {
 		c.JSON(400, gin.H{"error": "invalid body"})
 		return
 	}
-	targetIndex := -1
-	if body.Index != nil && *body.Index >= 0 && *body.Index < len(h.cfg.BedrockKey) {
-		targetIndex = *body.Index
-	}
-	if targetIndex == -1 && body.Match != nil {
-		match := strings.TrimSpace(*body.Match)
-		for i := range h.cfg.BedrockKey {
-			entry := h.cfg.BedrockKey[i]
-			if entry.APIKey == match || entry.AccessKeyID == match || entry.Name == match {
-				targetIndex = i
-				break
-			}
+	if err := providerSettingsService(h).PatchBedrockKey(body.Index, body.Match, *body.Value); err != nil {
+		if errors.Is(err, providersettings.ErrItemNotFound) {
+			c.JSON(404, gin.H{"error": "item not found"})
+			return
 		}
-	}
-	if targetIndex == -1 {
-		c.JSON(404, gin.H{"error": "item not found"})
-		return
-	}
-
-	entry := h.cfg.BedrockKey[targetIndex]
-	if body.Value.Name != nil {
-		entry.Name = strings.TrimSpace(*body.Value.Name)
-	}
-	if body.Value.Priority != nil {
-		entry.Priority = *body.Value.Priority
-	}
-	if body.Value.Prefix != nil {
-		entry.Prefix = strings.TrimSpace(*body.Value.Prefix)
-	}
-	if body.Value.AuthMode != nil {
-		entry.AuthMode = strings.TrimSpace(*body.Value.AuthMode)
-	}
-	if body.Value.APIKey != nil {
-		entry.APIKey = strings.TrimSpace(*body.Value.APIKey)
-	}
-	if body.Value.AccessKeyID != nil {
-		entry.AccessKeyID = strings.TrimSpace(*body.Value.AccessKeyID)
-	}
-	if body.Value.SecretAccessKey != nil {
-		entry.SecretAccessKey = strings.TrimSpace(*body.Value.SecretAccessKey)
-	}
-	if body.Value.SessionToken != nil {
-		entry.SessionToken = strings.TrimSpace(*body.Value.SessionToken)
-	}
-	if body.Value.Region != nil {
-		entry.Region = strings.TrimSpace(*body.Value.Region)
-	}
-	if body.Value.ForceGlobal != nil {
-		entry.ForceGlobal = *body.Value.ForceGlobal
-	}
-	if body.Value.BaseURL != nil {
-		entry.BaseURL = strings.TrimSpace(*body.Value.BaseURL)
-	}
-	if body.Value.ProxyURL != nil {
-		entry.ProxyURL = strings.TrimSpace(*body.Value.ProxyURL)
-	}
-	if body.Value.ProxyID != nil {
-		entry.ProxyID = strings.TrimSpace(*body.Value.ProxyID)
-	}
-	if body.Value.Models != nil {
-		entry.Models = append([]config.BedrockModel(nil), (*body.Value.Models)...)
-	}
-	if body.Value.Headers != nil {
-		entry.Headers = config.NormalizeHeaders(*body.Value.Headers)
-	}
-	if body.Value.ExcludedModels != nil {
-		entry.ExcludedModels = config.NormalizeExcludedModels(*body.Value.ExcludedModels)
-	}
-	normalizeBedrockKey(&entry)
-	prev := append([]config.BedrockKey(nil), h.cfg.BedrockKey...)
-	h.cfg.BedrockKey[targetIndex] = entry
-	h.cfg.SanitizeBedrockKeys()
-	if err := h.validateChannelNames(); err != nil {
-		h.cfg.BedrockKey = prev
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
@@ -782,15 +688,7 @@ func (h *Handler) PatchBedrockKey(c *gin.Context) {
 
 func (h *Handler) DeleteBedrockKey(c *gin.Context) {
 	if val := strings.TrimSpace(c.Query("api-key")); val != "" {
-		out := make([]config.BedrockKey, 0, len(h.cfg.BedrockKey))
-		for _, v := range h.cfg.BedrockKey {
-			if v.APIKey != val {
-				out = append(out, v)
-			}
-		}
-		if len(out) != len(h.cfg.BedrockKey) {
-			h.cfg.BedrockKey = out
-			h.cfg.SanitizeBedrockKeys()
+		if providerSettingsService(h).DeleteBedrockKeyByAPIKey(val) {
 			h.persist(c)
 			return
 		}
@@ -798,15 +696,7 @@ func (h *Handler) DeleteBedrockKey(c *gin.Context) {
 		return
 	}
 	if val := strings.TrimSpace(c.Query("access-key-id")); val != "" {
-		out := make([]config.BedrockKey, 0, len(h.cfg.BedrockKey))
-		for _, v := range h.cfg.BedrockKey {
-			if v.AccessKeyID != val {
-				out = append(out, v)
-			}
-		}
-		if len(out) != len(h.cfg.BedrockKey) {
-			h.cfg.BedrockKey = out
-			h.cfg.SanitizeBedrockKeys()
+		if providerSettingsService(h).DeleteBedrockKeyByAccessKeyID(val) {
 			h.persist(c)
 			return
 		}
@@ -814,15 +704,7 @@ func (h *Handler) DeleteBedrockKey(c *gin.Context) {
 		return
 	}
 	if val := strings.TrimSpace(c.Query("name")); val != "" {
-		out := make([]config.BedrockKey, 0, len(h.cfg.BedrockKey))
-		for _, v := range h.cfg.BedrockKey {
-			if v.Name != val {
-				out = append(out, v)
-			}
-		}
-		if len(out) != len(h.cfg.BedrockKey) {
-			h.cfg.BedrockKey = out
-			h.cfg.SanitizeBedrockKeys()
+		if providerSettingsService(h).DeleteBedrockKeyByName(val) {
 			h.persist(c)
 			return
 		}
@@ -831,9 +713,7 @@ func (h *Handler) DeleteBedrockKey(c *gin.Context) {
 	}
 	if idxStr := c.Query("index"); idxStr != "" {
 		var idx int
-		if _, err := fmt.Sscanf(idxStr, "%d", &idx); err == nil && idx >= 0 && idx < len(h.cfg.BedrockKey) {
-			h.cfg.BedrockKey = append(h.cfg.BedrockKey[:idx], h.cfg.BedrockKey[idx+1:]...)
-			h.cfg.SanitizeBedrockKeys()
+		if _, err := fmt.Sscanf(idxStr, "%d", &idx); err == nil && providerSettingsService(h).DeleteBedrockKeyByIndex(idx) {
 			h.persist(c)
 			return
 		}
@@ -843,7 +723,7 @@ func (h *Handler) DeleteBedrockKey(c *gin.Context) {
 
 // opencode-go-api-key: []OpenCodeGoKey
 func (h *Handler) GetOpenCodeGoKeys(c *gin.Context) {
-	c.JSON(200, gin.H{"opencode-go-api-key": normalizedOpenCodeGoKeyEntries(h.cfg.OpenCodeGoKey)})
+	c.JSON(200, gin.H{"opencode-go-api-key": providerSettingsService(h).OpenCodeGoKeys()})
 }
 
 func (h *Handler) PutOpenCodeGoKeys(c *gin.Context) {
@@ -863,18 +743,7 @@ func (h *Handler) PutOpenCodeGoKeys(c *gin.Context) {
 		}
 		arr = obj.Items
 	}
-	filtered := make([]config.OpenCodeGoKey, 0, len(arr))
-	for i := range arr {
-		normalizeOpenCodeGoKey(&arr[i])
-		if strings.TrimSpace(arr[i].APIKey) != "" {
-			filtered = append(filtered, arr[i])
-		}
-	}
-	prev := append([]config.OpenCodeGoKey(nil), h.cfg.OpenCodeGoKey...)
-	h.cfg.OpenCodeGoKey = filtered
-	h.cfg.SanitizeOpenCodeGoKeys()
-	if err := h.validateChannelNames(); err != nil {
-		h.cfg.OpenCodeGoKey = prev
+	if err := providerSettingsService(h).ReplaceOpenCodeGoKeys(arr); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
@@ -882,102 +751,21 @@ func (h *Handler) PutOpenCodeGoKeys(c *gin.Context) {
 }
 
 func (h *Handler) PatchOpenCodeGoKey(c *gin.Context) {
-	type openCodeGoPatch struct {
-		APIKey         *string            `json:"api-key"`
-		Name           *string            `json:"name"`
-		Priority       *int               `json:"priority"`
-		Prefix         *string            `json:"prefix"`
-		ProxyURL       *string            `json:"proxy-url"`
-		ProxyID        *string            `json:"proxy-id"`
-		Headers        *map[string]string `json:"headers"`
-		ExcludedModels *[]string          `json:"excluded-models"`
-		VisionFallback *string            `json:"vision-fallback-model"`
-		WorkspaceID    *string            `json:"workspace-id"`
-		AuthCookie     *string            `json:"auth-cookie"`
-	}
 	var body struct {
-		APIKey *string          `json:"api-key"`
-		Name   *string          `json:"name"`
-		Index  *int             `json:"index"`
-		Value  *openCodeGoPatch `json:"value"`
+		APIKey *string                           `json:"api-key"`
+		Name   *string                           `json:"name"`
+		Index  *int                              `json:"index"`
+		Value  *providersettings.OpenCodeGoPatch `json:"value"`
 	}
 	if err := c.ShouldBindJSON(&body); err != nil || body.Value == nil {
 		c.JSON(400, gin.H{"error": "invalid body"})
 		return
 	}
-	targetIndex := -1
-	if body.Index != nil && *body.Index >= 0 && *body.Index < len(h.cfg.OpenCodeGoKey) {
-		targetIndex = *body.Index
-	}
-	if targetIndex == -1 && body.APIKey != nil {
-		match := strings.TrimSpace(*body.APIKey)
-		for i := range h.cfg.OpenCodeGoKey {
-			if h.cfg.OpenCodeGoKey[i].APIKey == match {
-				targetIndex = i
-				break
-			}
+	if err := providerSettingsService(h).PatchOpenCodeGoKey(body.Index, body.APIKey, body.Name, *body.Value); err != nil {
+		if errors.Is(err, providersettings.ErrItemNotFound) {
+			c.JSON(404, gin.H{"error": "item not found"})
+			return
 		}
-	}
-	if targetIndex == -1 && body.Name != nil {
-		match := strings.TrimSpace(*body.Name)
-		for i := range h.cfg.OpenCodeGoKey {
-			if h.cfg.OpenCodeGoKey[i].Name == match {
-				targetIndex = i
-				break
-			}
-		}
-	}
-	if targetIndex == -1 {
-		c.JSON(404, gin.H{"error": "item not found"})
-		return
-	}
-
-	entry := h.cfg.OpenCodeGoKey[targetIndex]
-	if body.Value.APIKey != nil {
-		entry.APIKey = strings.TrimSpace(*body.Value.APIKey)
-	}
-	if body.Value.Name != nil {
-		entry.Name = strings.TrimSpace(*body.Value.Name)
-	}
-	if body.Value.Priority != nil {
-		entry.Priority = *body.Value.Priority
-	}
-	if body.Value.Prefix != nil {
-		entry.Prefix = strings.TrimSpace(*body.Value.Prefix)
-	}
-	if body.Value.ProxyURL != nil {
-		entry.ProxyURL = strings.TrimSpace(*body.Value.ProxyURL)
-	}
-	if body.Value.ProxyID != nil {
-		entry.ProxyID = strings.TrimSpace(*body.Value.ProxyID)
-	}
-	if body.Value.Headers != nil {
-		entry.Headers = config.NormalizeHeaders(*body.Value.Headers)
-	}
-	if body.Value.ExcludedModels != nil {
-		entry.ExcludedModels = config.NormalizeExcludedModels(*body.Value.ExcludedModels)
-	}
-	if body.Value.VisionFallback != nil {
-		entry.VisionFallbackModel = strings.TrimSpace(*body.Value.VisionFallback)
-	}
-	if body.Value.WorkspaceID != nil {
-		entry.WorkspaceID = strings.TrimSpace(*body.Value.WorkspaceID)
-	}
-	if body.Value.AuthCookie != nil {
-		entry.AuthCookie = strings.TrimSpace(*body.Value.AuthCookie)
-	}
-	normalizeOpenCodeGoKey(&entry)
-	if entry.APIKey == "" {
-		h.cfg.OpenCodeGoKey = append(h.cfg.OpenCodeGoKey[:targetIndex], h.cfg.OpenCodeGoKey[targetIndex+1:]...)
-		h.cfg.SanitizeOpenCodeGoKeys()
-		h.persist(c)
-		return
-	}
-	prev := append([]config.OpenCodeGoKey(nil), h.cfg.OpenCodeGoKey...)
-	h.cfg.OpenCodeGoKey[targetIndex] = entry
-	h.cfg.SanitizeOpenCodeGoKeys()
-	if err := h.validateChannelNames(); err != nil {
-		h.cfg.OpenCodeGoKey = prev
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
@@ -986,15 +774,7 @@ func (h *Handler) PatchOpenCodeGoKey(c *gin.Context) {
 
 func (h *Handler) DeleteOpenCodeGoKey(c *gin.Context) {
 	if apiKey := strings.TrimSpace(c.Query("api-key")); apiKey != "" {
-		out := make([]config.OpenCodeGoKey, 0, len(h.cfg.OpenCodeGoKey))
-		for _, v := range h.cfg.OpenCodeGoKey {
-			if v.APIKey != apiKey {
-				out = append(out, v)
-			}
-		}
-		if len(out) != len(h.cfg.OpenCodeGoKey) {
-			h.cfg.OpenCodeGoKey = out
-			h.cfg.SanitizeOpenCodeGoKeys()
+		if providerSettingsService(h).DeleteOpenCodeGoKeyByAPIKey(apiKey) {
 			h.persist(c)
 			return
 		}
@@ -1002,15 +782,7 @@ func (h *Handler) DeleteOpenCodeGoKey(c *gin.Context) {
 		return
 	}
 	if name := strings.TrimSpace(c.Query("name")); name != "" {
-		out := make([]config.OpenCodeGoKey, 0, len(h.cfg.OpenCodeGoKey))
-		for _, v := range h.cfg.OpenCodeGoKey {
-			if v.Name != name {
-				out = append(out, v)
-			}
-		}
-		if len(out) != len(h.cfg.OpenCodeGoKey) {
-			h.cfg.OpenCodeGoKey = out
-			h.cfg.SanitizeOpenCodeGoKeys()
+		if providerSettingsService(h).DeleteOpenCodeGoKeyByName(name) {
 			h.persist(c)
 			return
 		}
@@ -1019,9 +791,7 @@ func (h *Handler) DeleteOpenCodeGoKey(c *gin.Context) {
 	}
 	if idxStr := c.Query("index"); idxStr != "" {
 		var idx int
-		if _, err := fmt.Sscanf(idxStr, "%d", &idx); err == nil && idx >= 0 && idx < len(h.cfg.OpenCodeGoKey) {
-			h.cfg.OpenCodeGoKey = append(h.cfg.OpenCodeGoKey[:idx], h.cfg.OpenCodeGoKey[idx+1:]...)
-			h.cfg.SanitizeOpenCodeGoKeys()
+		if _, err := fmt.Sscanf(idxStr, "%d", &idx); err == nil && providerSettingsService(h).DeleteOpenCodeGoKeyByIndex(idx) {
 			h.persist(c)
 			return
 		}
@@ -1353,71 +1123,6 @@ func (h *Handler) DeleteCodexKey(c *gin.Context) {
 		}
 	}
 	c.JSON(400, gin.H{"error": "missing api-key or index"})
-}
-
-func normalizeOpenCodeGoKey(entry *config.OpenCodeGoKey) {
-	if entry == nil {
-		return
-	}
-	entry.Name = strings.TrimSpace(entry.Name)
-	entry.APIKey = strings.TrimSpace(entry.APIKey)
-	entry.Prefix = strings.TrimSpace(entry.Prefix)
-	entry.ProxyURL = strings.TrimSpace(entry.ProxyURL)
-	entry.ProxyID = strings.TrimSpace(entry.ProxyID)
-	entry.Headers = config.NormalizeHeaders(entry.Headers)
-	entry.ExcludedModels = config.NormalizeExcludedModels(entry.ExcludedModels)
-	entry.VisionFallbackModel = strings.TrimSpace(entry.VisionFallbackModel)
-	if workspaceID, err := normalizeOpenCodeGoWorkspaceID(entry.WorkspaceID); err == nil {
-		entry.WorkspaceID = workspaceID
-	} else {
-		entry.WorkspaceID = strings.TrimSpace(entry.WorkspaceID)
-	}
-	entry.AuthCookie = strings.TrimSpace(entry.AuthCookie)
-}
-
-func normalizedOpenCodeGoKeyEntries(entries []config.OpenCodeGoKey) []config.OpenCodeGoKey {
-	if len(entries) == 0 {
-		return nil
-	}
-	out := make([]config.OpenCodeGoKey, len(entries))
-	for i := range entries {
-		out[i] = entries[i]
-		normalizeOpenCodeGoKey(&out[i])
-	}
-	return out
-}
-
-func normalizeBedrockKey(entry *config.BedrockKey) {
-	if entry == nil {
-		return
-	}
-	entry.Name = strings.TrimSpace(entry.Name)
-	entry.Prefix = strings.TrimSpace(entry.Prefix)
-	entry.AuthMode = strings.TrimSpace(entry.AuthMode)
-	entry.APIKey = strings.TrimSpace(entry.APIKey)
-	entry.AccessKeyID = strings.TrimSpace(entry.AccessKeyID)
-	entry.SecretAccessKey = strings.TrimSpace(entry.SecretAccessKey)
-	entry.SessionToken = strings.TrimSpace(entry.SessionToken)
-	entry.Region = strings.TrimSpace(entry.Region)
-	entry.BaseURL = strings.TrimSpace(entry.BaseURL)
-	entry.ProxyURL = strings.TrimSpace(entry.ProxyURL)
-	entry.ProxyID = strings.TrimSpace(entry.ProxyID)
-	entry.Headers = config.NormalizeHeaders(entry.Headers)
-	entry.ExcludedModels = config.NormalizeExcludedModels(entry.ExcludedModels)
-	if len(entry.Models) == 0 {
-		return
-	}
-	normalized := make([]config.BedrockModel, 0, len(entry.Models))
-	for i := range entry.Models {
-		model := entry.Models[i]
-		model.Name = strings.TrimSpace(model.Name)
-		model.Alias = strings.TrimSpace(model.Alias)
-		if model.Name == "" && model.Alias == "" {
-			continue
-		}
-		normalized = append(normalized, model)
-	}
-	entry.Models = normalized
 }
 
 func ampSettingsService(h *Handler) *ampsettings.Service {
